requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-bootstrap4==23.1
django-extensions==3.2.1
djangorestframework==3.14.0
drf-yasg==1.21.5
factory-boy==3.2.1
Faker==18.7.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.2
